\name{TPAM-package}
\alias{TPAM-package}
\alias{TPAM}
\docType{package}
\title{
Classification using time-course gene expression data
}
\description{
Classification using time-course gene expression data
}
\details{
\tabular{ll}{
Package: \tab TPAM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-30\cr
License: \tab GPL\cr
LazyLoad: \tab false \cr
}

}
\author{
Yuping Zhang

Maintainer: Yuping Zhang <yupingz@stanford.edu>
}
\references{
Classification of patients from time-course gene expression. Biostatistics 2012}
\keyword{ Classification, Time Course }
