% --- Source file: man/tspc.internal.Rd ---
\name{superpc-internal}
\title{Internal superpc functions}
\alias{balanced.folds}  
\alias{cor.func}  
\alias{cox.coef.func}  
\alias{coxfunc}  
\alias{coxscor} 
\alias{coxstuff}  
\alias{coxvar}  
\alias{error.bars}  
\alias{jitter}  
\alias{lm.coef.func}
\alias{max.na}   
\alias{min.na}  
\alias{multiply.func}  
\alias{mysvd}  
\alias{ocoxvar}  
\alias{permute.rows} 
\alias{se.na} 
\alias{diag.disc}
\alias{lda_project}
\alias{my.nsccv}
\alias{my.pamr.cv}
\alias{my.pamr.cv.se}
\alias{my.pamr.plotcv}
\alias{nsc}
\alias{permute.rows}
\alias{pred.err}
\alias{safe.exp}
\alias{soft.shrink}
\alias{softmax}
\alias{sqrt_norm}
\alias{which.is.max}
\description{Internal TPAM functions}
\author{Yuping Zhang}
\details{
  These are not to be called by the user.
}
\keyword{internal}
